# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "在东软的教学包基础上去做拓展",
      "title": "饿了么项目复刻",
      "termsOfService": "http://swagger.io/terms/",
      "contact": {
        "name": "跑路了",
        "url": "http://www.swagger.io/support",
        "email": "support@swagger.io"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      },
      "version": "1.0"
    },
  "host": "localhost:9090",
  "basePath": "/api/v1",
  "paths": {
    "/api/v1/comments/orders/{order_id}": {
      "post": {
        "description": "用户为指定id的订单添加评价",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "评价管理"
        ],
        "summary": "用户评价订单",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/orders/{order_id}": {
      "get": {
        "description": "获取指定id的订单详情",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "订单管理"
        ],
        "summary": "获取订单详情",
        "parameters": [
          {
            "type": "integer",
            "description": "订单id",
            "name": "order_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OrderDetailResponse",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "ErrorResponse",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "description": "取消指定id的订单",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "订单管理"
        ],
        "summary": "取消订单",
        "parameters": [
          {
            "type": "integer",
            "description": "订单id",
            "name": "order_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "succeed",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "ErrorResponse",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/riders/orders/{order_id}/accept": {
      "post": {
        "description": "骑手接收指定id的订单",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "骑手管理"
        ],
        "summary": "骑手接单",
        "parameters": [
          {
            "type": "integer",
            "description": "订单id",
            "name": "order_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "接单成功",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "ErrorResponse",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/riders/orders/{order_id}/complete": {
      "post": {
        "description": "骑手完成指定id的订单",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "骑手管理"
        ],
        "summary": "骑手完成订单",
        "parameters": [
          {
            "type": "integer",
            "description": "订单id",
            "name": "order_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "订单完成",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "ErrorResponse",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/dish/add": {
      "post": {
        "description": "添加菜品",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "菜品管理"
        ],
        "summary": "添加菜品",
        "parameters": [
          {
            "type": "string",
            "description": "菜品名称",
            "name": "name",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "菜品描述",
            "name": "description",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "菜品图片",
            "name": "picture",
            "in": "formData"
          },
          {
            "type": "number",
            "description": "菜品价格",
            "name": "price",
            "in": "formData",
            "required": true
          },
          {
            "type": "integer",
            "description": "所属餐厅id",
            "name": "merchantID",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "添加成功",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "输入非法",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "添加失败",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/dish/fuzzy": {
      "post": {
        "description": "根据菜品名称模糊搜索菜品信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "菜品管理"
        ],
        "summary": "模糊搜索菜品信息",
        "parameters": [
          {
            "type": "string",
            "description": "菜品名称",
            "name": "name",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Dish",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.Dish"
              }
            }
          },
          "400": {
            "description": "请求参数不能为空",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "请求资源不存在",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "查询失败",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/dish/list": {
      "get": {
        "description": "获取所有菜品列表",
        "produces": [
          "application/json"
        ],
        "tags": [
          "菜品管理"
        ],
        "summary": "列出所有菜品",
        "responses": {
          "200": {
            "description": "Dish",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.Dish"
              }
            }
          },
          "404": {
            "description": "资源未找到",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "查询失败",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/dish/perfect": {
      "post": {
        "description": "根据菜品名称准确获取菜品信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "菜品管理"
        ],
        "summary": "准确获取菜品信息",
        "parameters": [
          {
            "type": "string",
            "description": "菜品名称",
            "name": "name",
            "in": "formData",
            "required": true
          },
          {
            "type": "integer",
            "description": "所属餐厅id",
            "name": "merchantID",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Dish",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.Dish"
              }
            }
          },
          "400": {
            "description": "输入参数不能为空",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "请求资源不存在",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "查询失败",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/merchant": {
      "delete": {
        "description": "根据 id 删除指定商家",
        "produces": [
          "application/json"
        ],
        "tags": [
          "商家管理"
        ],
        "summary": "根据 id 删除指定商家",
        "parameters": [
          {
            "type": "integer",
            "description": "商家id",
            "name": "id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "删除失败",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/merchant/add": {
      "post": {
        "description": "添加商家",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "商家管理"
        ],
        "summary": "添加商家",
        "parameters": [
          {
            "type": "string",
            "description": "餐厅名称",
            "name": "name",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "餐厅地址",
            "name": "address",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "餐厅电话",
            "name": "phone",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "添加成功",
            "schema": {
              "$ref": "#/definitions/models.Merchant"
            }
          },
          "400": {
            "description": "添加失败",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "输入非法",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/merchant/fuzzy": {
      "get": {
        "description": "根据商家名称模糊搜索商家信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "商家管理"
        ],
        "summary": "模糊搜索商家信息",
        "parameters": [
          {
            "type": "string",
            "description": "商家名称",
            "name": "name",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.Merchant"
              }
            }
          },
          "400": {
            "description": "输入非法",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "资源未找到",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "查询失败",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/merchant/list": {
      "get": {
        "description": "获取所有商家列表",
        "produces": [
          "application/json"
        ],
        "tags": [
          "商家管理"
        ],
        "summary": "列出所有商家",
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.Merchant"
              }
            }
          },
          "404": {
            "description": "资源未找到",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "查询失败",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/merchant/perfect": {
      "post": {
        "description": "根据商家名称获取商家信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "商家管理"
        ],
        "summary": "准确获取商家信息",
        "parameters": [
          {
            "type": "string",
            "description": "商家名称",
            "name": "name",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.Merchant"
              }
            }
          },
          "400": {
            "description": "输入非法",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "资源未找到",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "查询失败",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/orders": {
      "get": {
        "description": "获取所有订单列表",
        "produces": [
          "application/json"
        ],
        "summary": "列出所有订单",
        "responses": {
          "200": {
            "description": "succeed",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "ErrorResponse",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "post": {
        "description": "用户创建新订单",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "创建订单",
        "parameters": [
          {
            "description": "新订单信息",
            "name": "order",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "ErrorResponse",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "ErrorResponse",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/users": {
      "post": {
        "description": "创建新用户",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "创建用户",
        "parameters": [
          {
            "description": "新用户信息",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "ErrorResponse",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "ErrorResponse",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "description": "用户使用用户名和密码登录",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "summary": "用户登录",
        "parameters": [
          {
            "description": "用户登录信息",
            "name": "login",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "ErrorResponse",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "ErrorResponse",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "description": "用户登出",
        "produces": [
          "application/json"
        ],
        "summary": "用户登出",
        "responses": {
          "204": {
            "description": "退出成功",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "ErrorResponse",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "models.Comment": {
      "type": "object",
      "properties": {
        "content": {
          "description": "评论内容",
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "deletedAt": {
          "type": "object",
          "properties": {
            "time": {
              "type": "string"
            },
            "valid": {
              "description": "Valid is true if Time is not NULL",
              "type": "boolean"
            }
          }
        },
        "dishID": {
          "description": "评论所属菜品id",
          "type": "integer"
        },
        "id": {
          "type": "integer"
        },
        "score": {
          "description": "评论评分",
          "type": "integer"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "models.Dish": {
      "type": "object",
      "required": [
        "description",
        "merchantID",
        "name",
        "picture",
        "price"
      ],
      "properties": {
        "comments": {
          "description": "一个菜品有很多评价，使用外键关联",
          "type": "array",
          "items": {
            "$ref": "#/definitions/models.Comment"
          }
        },
        "createdAt": {
          "type": "string"
        },
        "deletedAt": {
          "type": "object",
          "properties": {
            "time": {
              "type": "string"
            },
            "valid": {
              "description": "Valid is true if Time is not NULL",
              "type": "boolean"
            }
          }
        },
        "description": {
          "description": "菜品描述",
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "merchantID": {
          "description": "所属餐厅id",
          "type": "integer"
        },
        "name": {
          "description": "菜品名称",
          "type": "string"
        },
        "picture": {
          "description": "菜品图片",
          "type": "string"
        },
        "price": {
          "description": "菜品价格",
          "type": "number"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "models.Merchant": {
      "type": "object",
      "required": [
        "address",
        "name",
        "phone"
      ],
      "properties": {
        "address": {
          "description": "餐厅地址",
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "deletedAt": {
          "type": "object",
          "properties": {
            "time": {
              "type": "string"
            },
            "valid": {
              "description": "Valid is true if Time is not NULL",
              "type": "boolean"
            }
          }
        },
        "dishes": {
          "description": "一个餐厅有多个菜品，使用外键关联",
          "type": "array",
          "items": {
            "$ref": "#/definitions/models.Dish"
          }
        },
        "id": {
          "type": "integer"
        },
        "name": {
          "description": "餐厅名称",
          "type": "string"
        },
        "phone": {
          "description": "餐厅电话",
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    }
  }
}